Round to the nearest int

+++ components/disk.c
@@ -27,8 +27,8 @@ disk_perc(const char *path)
 		return NULL;
 	}
 
-	return bprintf("%d", (int)(100 *
-	               (1 - ((double)fs.f_bavail / (double)fs.f_blocks))));
+	return bprintf("%.0f", 100 *
+	               (1 - (double)fs.f_bavail / (double)fs.f_blocks));
 }
 
 const char *
